requirements.txt
flask==3.1.2
pillow==12.0.0
google-generativeai==0.8.5
openai==0.27.8
psycopg2-binary>=2.9
python-dotenv>=1.0.0